Enable use of Jacoco instead of EclEmma for code coverage.
The following enables Jacoco to be used for code coverage and report
generation.

* Remove the eclemma agent.
* Remove the code coverage profile for emma.
* Created a profile for jacoco.
* Enable generation of HTML and XML reports or code coverage.

This uses a mixture of Mavena and Ant to get the job done, as currently
the report goal for he jacoco-maven-plugin does not correctly report
instrumented classes outside of the test project.  You have to specifically
tell it where the source files are and since the maven plugin doesn't look
outside it's directory we have to use the ant plugin to find the sources.

Output is generated in target/jacoco and target/jacoco/report.  The html
reports can be made availble through jenkins.  Working on an updated
XSLT to convert jacoco.xml to either emma or cobertura xml format for use
with jenkins/hudson plugins.